# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TraceCrypt Unified \u2014 Evidence-led Attribution",
    "description": "Read-only investigative triage. Does not identify persons, establish guilt, guarantee recovery, or execute freezes.",
    "version": "unified-2.0"
  },
  "paths": {
...
    "/api/v1/operations/intakes": {
      "post": {
        "tags": [
          "operations-automation"
        ],
        "summary": "Intake Manual",
        "operationId": "intake_manual_api_v1_operations_intakes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplaintPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/integrations/complaints": {
      "post": {
        "tags": [
          "operations-automation"
        ],
        "summary": "Intake Gateway",
        "operationId": "intake_gateway_api_v1_integrations_complaints_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplaintPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/operations/jobs": {
      "get": {
        "tags": [
          "operations-automation"
        ],
        "summary": "Jobs",
        "operationId": "jobs_api_v1_operations_jobs_get",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/operations/jobs/{job_id}": {
      "get": {
        "tags": [
          "operations-automation"
        ],
        "summary": "Job Detail",
        "operationId": "job_detail_api_v1_operations_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/operations/jobs/{job_id}/retry": {
      "post": {
        "tags": [
          "operations-automation"
        ],
        "summary": "Retry",
        "operationId": "retry_api_v1_operations_jobs__job_id__retry_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/operations/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "operations-automation"
        ],
        "summary": "Cancel",
        "operationId": "cancel_api_v1_operations_jobs__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/operations/signals": {
      "get": {
        "tags": [
          "operations-automation"
        ],
        "summary": "Signals",
        "operationId": "signals_api_v1_operations_signals_get",
        "parameters": [
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "latest",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Latest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/operations/signals/{signal_id}/acknowledge": {
      "post": {
        "tags": [
          "operations-automation"
        ],
        "summary": "Acknowledge",
        "operationId": "acknowledge_api_v1_operations_signals__signal_id__acknowledge_post",
        "parameters": [
          {
            "name": "signal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Signal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cases/{case_id}/investigations/{run_id}/attribution": {
      "get": {
        "tags": [
          "operations-automation"
        ],
        "summary": "Attribution",
        "operationId": "attribution_api_v1_cases__case_id__investigations__run_id__attribution_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComplaintPayload": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "external_reference": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9._:/-]+$",
            "title": "External Reference"
          },
          "title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Title"
          },
          "allegation_type": {
            "type": "string",
            "enum": [
              "investment_scam",
              "task_fraud",
              "sextortion",
              "ransomware",
              "phishing",
              "darknet",
              "organized_financial_crime",
              "other"
            ],
            "title": "Allegation Type",
            "default": "other"
          },
          "wallets": {
            "items": {
              "$ref": "#/components/schemas/WalletReport"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Wallets"
          },
          "incident_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Incident Start"
          },
          "analysis_cutoff": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Cutoff"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "external_reference",
          "title",
          "wallets"
        ],
        "title": "ComplaintPayload"
      },
...
    }
  }
}